requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMuPDF==1.24.14
PySimpleGUI==4.70.1
